Test in Firefox instead of Chrome
Chrome 87 just shipped the CookieStore implementation and it's not
possible to null out the `window.cookieStore` variable. So we're
switching to test in Firefox which hasn't implemented the feature yet so
that we can continue to test.

Not a great solution but maybe we can find a better one in the future,
this should do for now.